+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 * [api]
   - Fix support for multilines starting from empty line(s) (without multiline marker)
   - Change dry-run config report to better identify it and remove additional new lines (show exact result)
-  - Create parent directories when config is created (not existing config case)  
+  - Create parent directories when config is created (not existing config case)
+  - Re-style configuration not changed report to look like other reports.
+  - Do not show warning when config not changes
 
 ### 1.2.0 (2021-10-09)
 * [api]