+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 1.3.0 (2021-10-25)
 * [api]
   - Fix support for multilines starting from empty line(s) (without multiline marker)
   - Change dry-run config report to better identify it and remove additional new lines (show exact result)